Fix a timeout issue in RRQ with windowsize>1
Previously, if the connection timeouts during an RRQ request with a
windowsize>1, there will still be some retransmits in self.retransmits
which will try to use the now closed connection, fail, and raise an
error.
A test was added to rfc7440_test.py that reproduces this error.
protocols.py was changed to fix this issue.