Server should respond over same socker UDP socket.
When running in docker (--net=bridge) on Ubuntu 22.04
we hit a problem where the TFTP responses were not
routed properly. After some investigation we concluded
that the response got lost because it was coming from a
different (random) host:port.

The patch in question re-uses the server socket for
the UDP response so that reverse NAT treversal is
possible in stricter network topologies.